requirements.txt
Pillow
bs4
highlight_text
matplotlib
mplsoccer
pandas
requests
retrying
seaborn
selenium
sqlalchemy
unidecode
utils
tabulate
urllib3
shapely
scikit-learn
psycopg2-binary
webdriver_manager
google-cloud-storage
selenium
pytz
PyYAML
pyarrow